+++ forest/bridge/forestbridge.h
@@ -123,7 +123,7 @@ struct ForestBridge {
    */
   void dump(vector<vector<unsigned int> >& predTree,
             vector<vector<double> >& splitTree,
-            vector<vector<double> >& lhDelTree,
+            vector<vector<size_t> >& lhDelTree,
             vector<vector<unsigned char> >& facSplitTree) const;
   
 private:
