+++ forest/bridge/forestrw.h
@@ -26,7 +26,7 @@ struct ForestRW {
   static void dump(const class Forest* forest,
 		   vector<vector<unsigned int> >& predTree,
 		   vector<vector<double> >& splitTree,
-		   vector<vector<double> >& lhDelTree,
+		   vector<vector<size_t> >& lhDelTree,
 		   vector<vector<unsigned char> >& facSplitTree);
 };
